Revert "mv model_saver to __init__()"
This reverts commit 0d7e595fc79dab16c96d8df75c428a0aa50050d1.
